requirements.txt
discord.py
python-dotenv
apscheduler
firebase
python-firebase
python_jwt
gcloud
sseclient
pycryptodome
requests-toolbelt
numpy